+++ tests/model.rs
@@ -16,9 +16,9 @@ fn test_digital_asset_exchange() {
 
 #[cfg(feature = "reader-xml")]
 #[test]
-fn test_extensible_3d_graphics() {
+fn test_extensible3d_graphics() {
     let format = FileFormat::from_file("fixtures/model/sample.x3d").unwrap();
-    assert_eq!(format, FileFormat::Extensible3DGraphics);
+    assert_eq!(format, FileFormat::Extensible3dGraphics);
 }
 
 #[test]
@@ -34,13 +34,55 @@ fn test_google_draco() {
 }
 
 #[test]
-fn test_polygon_file_format() {
-    let format = FileFormat::from_file("fixtures/model/sample.ply").unwrap();
-    assert_eq!(format, FileFormat::PolygonFileFormat);
+fn test_inter_quake_export() {
+    let format = FileFormat::from_file("fixtures/model/sample.iqe").unwrap();
+    assert_eq!(format, FileFormat::InterQuakeExport);
 }
 
 #[test]
-fn test_stereolithography() {
-    let format = FileFormat::from_file("fixtures/model/sample.stl").unwrap();
-    assert_eq!(format, FileFormat::Stereolithography);
+fn test_inter_quake_model() {
+    let format = FileFormat::from_file("fixtures/model/sample.iqm").unwrap();
+    assert_eq!(format, FileFormat::InterQuakeModel);
+}
+
+#[test]
+fn test_magicavoxel() {
+    let format = FileFormat::from_file("fixtures/model/sample.vox").unwrap();
+    assert_eq!(format, FileFormat::Magicavoxel);
+}
+
+#[test]
+fn test_model3d_binary() {
+    let format = FileFormat::from_file("fixtures/model/sample.m3d").unwrap();
+    assert_eq!(format, FileFormat::Model3dBinary);
+}
+
+#[test]
+fn test_polygon_ascii() {
+    let format = FileFormat::from_file("fixtures/model/sample1.ply").unwrap();
+    assert_eq!(format, FileFormat::PolygonAscii);
+}
+
+#[test]
+fn test_polygon_binary() {
+    let format = FileFormat::from_file("fixtures/model/sample2.ply").unwrap();
+    assert_eq!(format, FileFormat::PolygonBinary);
+}
+
+#[test]
+fn test_stereolithography_ascii() {
+    let format = FileFormat::from_file("fixtures/model/sample1.stl").unwrap();
+    assert_eq!(format, FileFormat::StereolithographyAscii);
+}
+
+#[test]
+fn test_stereolithography_binary() {
+    let format = FileFormat::from_file("fixtures/model/sample2.stl").unwrap();
+    assert_eq!(format, FileFormat::StereolithographyBinary);
+}
+
+#[test]
+fn test_universal3d() {
+    let format = FileFormat::from_file("fixtures/model/sample.u3d").unwrap();
+    assert_eq!(format, FileFormat::Universal3d);
 }