+++ tests/text.rs
@@ -37,6 +37,12 @@ fn test_lua_script() {
 }
 
 #[test]
+fn test_model3d_ascii() {
+    let format = FileFormat::from_file("fixtures/text/sample.a3d").unwrap();
+    assert_eq!(format, FileFormat::Model3dAscii);
+}
+
+#[test]
 fn test_ms_dos_batch() {
     let format = FileFormat::from_file("fixtures/text/sample.bat").unwrap();
     assert_eq!(format, FileFormat::MsDosBatch);
@@ -89,4 +95,10 @@ fn test_vcalendar() {
 fn test_vcard() {
     let format = FileFormat::from_file("fixtures/text/sample.vcf").unwrap();
     assert_eq!(format, FileFormat::Vcard);
+}
+
+#[test]
+fn test_webassembly_text() {
+    let format = FileFormat::from_file("fixtures/text/sample.wat").unwrap();
+    assert_eq!(format, FileFormat::WebassemblyText);
 }